chore(components): rename tooltip, truncate
