+++ packages/components/src/components/Tooltip/all.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Tooltip } from './index';
+import { Tooltip } from './Tooltip';
 import { storiesOf } from '@storybook/react';
 
 const TooltipWrapper = styled.div`
