+++ packages/components/src/index.ts
@@ -39,9 +39,9 @@ export * from './components/loaders/Loader';
 export * from './components/loaders/Progress';
 export * from './components/loaders/FluidSpinner';
 export * from './components/loaders/LoadingContent';
-export * from './components/Tooltip';
+export * from './components/Tooltip/Tooltip';
 export * from './components/Timerange/Timerange';
-export * from './components/Truncate';
+export * from './components/Truncate/Truncate';
 export * from './components/form/SelectBar';
 export * from './components/HoverAnimation';
 export * from './components/Fade';
